feat: 스토리 데이터 및 UI 업데이트
- StoryData에 createdAt, likeCount 추가
- StoreInfoData에 favoriteCount를 Int에서 Long으로 변경
- Story UI에 좋아요 갯수, 설명을 포함한 세부 정보 표시 추가
- Story UI에 Story를 클릭시 자세한 내용을 볼 수 있는 다이얼로그 추가

+++ app/src/main/java/com/store_me/storeme/data/StoreInfoData.kt
@@ -30,7 +30,7 @@ data class StoreInfoData(
     val customCategory: String,
     val location: String,
     val locationCode: Int,
-    val favoriteCount: Int,
+    val favoriteCount: Long,
 )
 
 /**
